Collapse hot-path byte chains into word compares
Match true, false, and null literals with one unaligned 32-bit load
against a constant image instead of three or four byte compares, in
the validator, the tape builder, the cursor, and the dynamic parser.
Check each object key's closing quote and colon with one 16-bit load
in the cursor and the expected-key matcher. Copy packed field names
through base pointers so the encoder's block copy drops its bounds
checks while staying inlined, and give pointer, map, bytes, quoted,
interface, and any decoders an inlined not-null pretest so present
values skip the TryNull call. skipSpace guards with an unsigned
compare so its byte load stops re-checking bounds.

BuildIndex -12%, BuildIndexLarge -8.5%, ValidNumber -5.4%, EncodeMap
-8.4%, with small broad wins across decode, parse, and marshal and no
regressions; an integer-side digit probe that traded short-number cost
for long-run speed measured as a net loss and was not kept.

+++ space.go
@@ -8,7 +8,10 @@ import "encoding/binary"
 // budget is 80 and one call to a non-inlineable function costs 57 by
 // itself, so almost any addition here de-inlines every call site.
 func skipSpace(src []byte, i int) int {
-	for i < len(src) {
+	// The unsigned form of the guard lets the prover drop the bounds check
+	// on src[i]: i is never negative here, and if it ever were the loop
+	// would simply exit as it does today.
+	for uint(i) < uint(len(src)) {
 		c := src[i]
 		if c > ' ' {
 			return i
